refactor(v3): remove semantic aliases from registration
